requirements.txt
pyperclip==1.8.2
python-gnupg==0.5.6
Flask==2.3.2
Werkzeug==2.3.6
